drivers/iio/frequency/cf_axi_dds: Add support for AD9162

+++ drivers/iio/frequency/ad9739a.c
@@ -201,7 +201,7 @@ static int ad9739a_prepare(struct cf_axi_converter *conv)
 	}
 }
 
-static unsigned long ad9739a_get_data_clk(struct cf_axi_converter *conv)
+static unsigned long long ad9739a_get_data_clk(struct cf_axi_converter *conv)
 {
 	return clk_get_rate(conv->clk[CLK_DAC]);
 }
